backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.35

# Stock Data
yfinance==0.2.43
pykrx==1.0.45

# News
newsapi-python==0.2.7
feedparser==6.0.11

# AI
anthropic==0.34.2

# Scheduler
apscheduler==3.10.4

# Utils
python-dotenv==1.0.1
httpx==0.27.2
pandas==2.2.3
setuptools==69.5.1    # 70+ 에서 pkg_resources top-level 제거됨 — pykrx 호환 버전 고정
